Fix Docker build paths and clean modules
