+++ Dockerfile
@@ -8,9 +8,9 @@ RUN apt-get update && apt-get install -y curl gnupg && \
     curl -fsSL https://deb.nodesource.com/setup_18.x | bash - && \
     apt-get install -y nodejs && \
     apt-get clean && rm -rf /var/lib/apt/lists/*
-
-# Nur die minimal notwendigen Python-Pakete (entspricht Projektvorgaben)
-RUN pip install --no-cache-dir flask jinja2 requests pyyaml werkzeug
+# Python-Abhängigkeiten installieren
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
 
 # ---- Frontend Build Stage ----------------------------------------------------
 FROM base AS frontend_build
@@ -25,10 +25,11 @@ RUN npm run build
 # ---- Controller Stage (liefert Frontend aus) ---------------------------------
 FROM base AS controller_stage
 WORKDIR /app
-# Backend/Controller-Code
-COPY controller.py /app/controller.py
-# Falls mehr Backenddateien existieren (Blueprints/Module), hier mitkopieren:
-# COPY controller/ /app/controller/
+# Backend/Controller-Code und gemeinsame Module
+COPY controller/ /app/controller/
+COPY src/ /app/src/
+COPY common/ /app/common/
+COPY config.json /app/config.json
 
 # Frontend-Build ins Image legen
 COPY --from=frontend_build /app/frontend/dist /app/frontend/dist
@@ -39,21 +40,18 @@ RUN pip install --no-cache-dir gunicorn
 ENV FRONTEND_DIST=/app/frontend/dist \
     LOG_LEVEL=INFO
 EXPOSE 8081
-# WSGI-Entry "controller:app" anpassen, falls dein App-Objekt anders heißt/liegt
-CMD ["gunicorn", "-b", "0.0.0.0:8081", "controller:app"]
+# WSGI-Entry zeigt auf controller/controller.py
+CMD ["gunicorn", "-b", "0.0.0.0:8081", "controller.controller:app"]
 
 # ---- AI-Agent Stage ----------------------------------------------------------
 FROM base AS ai_agent_stage
 WORKDIR /app
-# Agent-Code
-# vorher (fehlerhaft, wenn ai_agent.py in src/ liegt)
-# COPY ai_agent.py /app/ai_agent.py
-
-# AI-Agent-Dateien kopieren
-COPY agent/ai_agent.py /app/ai_agent.py
-
-# Weitere Module kopieren
-# COPY src/ /app/src/
+# Agent und gemeinsame Module
+COPY agent/ /app/agent/
+COPY src/ /app/src/
+COPY common/ /app/common/
+COPY config.json /app/config.json
 ENV LOG_LEVEL=INFO
 EXPOSE 5000
-CMD ["python", "-u", "ai_agent.py"]
+CMD ["python", "-u", "agent/ai_agent.py"]
+
