+++ common/http_client.py
@@ -1,29 +1,21 @@
+"""Small HTTP helper functions with retry support."""
+
+from __future__ import annotations
+
 import json
-import json
+import logging
 import time
 import urllib.error
 import urllib.parse
 import urllib.request
-import logging
 from typing import Any, Dict
 
 logger = logging.getLogger(__name__)
 
 
 def http_get(url: str, retries: int = 5, delay: float = 1.0, timeout: float = 10.0) -> Any:
-    """Perform a GET request with retry and basic JSON decoding.
+    """Perform a GET request with retries and JSON decoding."""
 
-    Parameters
-    ----------
-    url: str
-        Target URL.
-    retries: int
-        Number of attempts before giving up.
-    delay: float
-        Delay in seconds between retries.
-    timeout: float
-        Timeout in seconds passed to ``urllib.request.urlopen``.
-    """
     last_err: Exception | None = None
     for attempt in range(1, retries + 1):
         try:
@@ -41,7 +33,8 @@ def http_get(url: str, retries: int = 5, delay: float = 1.0, timeout: float = 10
                 )
                 time.sleep(delay)
             else:
-                raise last_err
+                logger.error("[http_get] failed after %s attempts: %s", retries, e)
+                return None
 
 
 def http_post(
@@ -54,7 +47,8 @@ def http_post(
     delay: float = 1.0,
     timeout: float = 10.0,
 ) -> Any:
-    """Perform a POST request with retry and basic JSON decoding."""
+    """Perform a POST request with retries and JSON decoding."""
+
     last_err: Exception | None = None
     for attempt in range(1, retries + 1):
         try:
@@ -81,65 +75,6 @@ def http_post(
                 )
                 time.sleep(delay)
             else:
-                raise last_err
-import json
-import logging
-import time
-from typing import Any, Dict, Optional
+                logger.error("[http_post] failed after %s attempts: %s", retries, e)
+                return None
 
-import requests
-from requests.exceptions import RequestException
-
-logger = logging.getLogger(__name__)
-
-def http_get(url: str, retries: int = 3, delay: float = 1.0) -> Any:
-    """
-    HTTP GET-Anfrage mit Wiederholungsversuchen bei Fehlern.
-
-    Args:
-        url: Die URL für die GET-Anfrage
-        retries: Anzahl der Wiederholungsversuche
-        delay: Verzögerung zwischen den Versuchen in Sekunden
-
-    Returns:
-        Der deserialisierte JSON-Inhalt oder None bei Fehlern
-    """
-    for attempt in range(1, retries + 1):
-        try:
-            response = requests.get(url, timeout=10)
-            response.raise_for_status()
-            return response.json()
-        except RequestException as e:
-            logger.warning("HTTP GET %s fehlgeschlagen (Versuch %d/%d): %s", 
-                          url, attempt, retries, e)
-            if attempt < retries:
-                time.sleep(delay)
-    logger.error("HTTP GET %s nach %d Versuchen fehlgeschlagen", url, retries)
-    return None
-
-def http_post(url: str, data: Dict[str, Any], 
-              retries: int = 3, delay: float = 1.0) -> Optional[Any]:
-    """
-    HTTP POST-Anfrage mit Wiederholungsversuchen bei Fehlern.
-
-    Args:
-        url: Die URL für die POST-Anfrage
-        data: Die zu sendenden Daten
-        retries: Anzahl der Wiederholungsversuche
-        delay: Verzögerung zwischen den Versuchen in Sekunden
-
-    Returns:
-        Der deserialisierte JSON-Inhalt oder None bei Fehlern
-    """
-    for attempt in range(1, retries + 1):
-        try:
-            response = requests.post(url, json=data, timeout=10)
-            response.raise_for_status()
-            return response.json() if response.content else None
-        except RequestException as e:
-            logger.warning("HTTP POST %s fehlgeschlagen (Versuch %d/%d): %s", 
-                          url, attempt, retries, e)
-            if attempt < retries:
-                time.sleep(delay)
-    logger.error("HTTP POST %s nach %d Versuchen fehlgeschlagen", url, retries)
-    return None
